Move asset types into their own go files
