+++ assets/references.go
@@ -3,17 +3,9 @@ package assets
 import (
 	"fmt"
 
-	validator "gopkg.in/go-playground/validator.v9"
-
 	"github.com/nyaruka/gocommon/jsonx"
 	"github.com/nyaruka/gocommon/uuids"
-	"github.com/nyaruka/goflow/utils"
 )
-
-func init() {
-	utils.RegisterStructValidator(GroupReferenceValidation, GroupReference{})
-	utils.RegisterStructValidator(LabelReferenceValidation, LabelReference{})
-}
 
 // Reference is interface for all reference types
 type Reference interface {
@@ -30,434 +22,6 @@ type UUIDReference interface {
 	GenericUUID() uuids.UUID
 }
 
-// ChannelReference is used to reference a channel
-type ChannelReference struct {
-	UUID ChannelUUID `json:"uuid" validate:"required,uuid"`
-	Name string      `json:"name"`
-}
-
-// NewChannelReference creates a new channel reference with the given UUID and name
-func NewChannelReference(uuid ChannelUUID, name string) *ChannelReference {
-	return &ChannelReference{UUID: uuid, Name: name}
-}
-
-// Type returns the name of the asset type
-func (r *ChannelReference) Type() string {
-	return "channel"
-}
-
-// GenericUUID returns the untyped UUID
-func (r *ChannelReference) GenericUUID() uuids.UUID {
-	return uuids.UUID(r.UUID)
-}
-
-// Identity returns the unique identity of the asset
-func (r *ChannelReference) Identity() string {
-	return string(r.UUID)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *ChannelReference) Variable() bool {
-	return false
-}
-
-func (r *ChannelReference) String() string {
-	return fmt.Sprintf("%s[uuid=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ UUIDReference = (*ChannelReference)(nil)
-
-// ClassifierReference is used to reference a classifier
-type ClassifierReference struct {
-	UUID ClassifierUUID `json:"uuid" validate:"required,uuid"`
-	Name string         `json:"name"`
-}
-
-// NewClassifierReference creates a new classifier reference with the given UUID and name
-func NewClassifierReference(uuid ClassifierUUID, name string) *ClassifierReference {
-	return &ClassifierReference{UUID: uuid, Name: name}
-}
-
-// Type returns the name of the asset type
-func (r *ClassifierReference) Type() string {
-	return "classifier"
-}
-
-// GenericUUID returns the untyped UUID
-func (r *ClassifierReference) GenericUUID() uuids.UUID {
-	return uuids.UUID(r.UUID)
-}
-
-// Identity returns the unique identity of the asset
-func (r *ClassifierReference) Identity() string {
-	return string(r.UUID)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *ClassifierReference) Variable() bool {
-	return false
-}
-
-func (r *ClassifierReference) String() string {
-	return fmt.Sprintf("%s[uuid=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ UUIDReference = (*ClassifierReference)(nil)
-
-// GroupReference is used to reference a group
-type GroupReference struct {
-	UUID      GroupUUID `json:"uuid,omitempty" validate:"omitempty,uuid4"`
-	Name      string    `json:"name,omitempty"`
-	NameMatch string    `json:"name_match,omitempty" engine:"evaluated"`
-}
-
-// NewGroupReference creates a new group reference with the given UUID and name
-func NewGroupReference(uuid GroupUUID, name string) *GroupReference {
-	return &GroupReference{UUID: uuid, Name: name}
-}
-
-// NewVariableGroupReference creates a new group reference from the given templatized name match
-func NewVariableGroupReference(nameMatch string) *GroupReference {
-	return &GroupReference{NameMatch: nameMatch}
-}
-
-// Type returns the name of the asset type
-func (r *GroupReference) Type() string {
-	return "group"
-}
-
-// GenericUUID returns the untyped UUID
-func (r *GroupReference) GenericUUID() uuids.UUID {
-	return uuids.UUID(r.UUID)
-}
-
-// Identity returns the unique identity of the asset
-func (r *GroupReference) Identity() string {
-	return string(r.UUID)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *GroupReference) Variable() bool {
-	return r.Identity() == ""
-}
-
-func (r *GroupReference) String() string {
-	return fmt.Sprintf("%s[uuid=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ UUIDReference = (*GroupReference)(nil)
-
-// FieldReference is a reference to a field
-type FieldReference struct {
-	Key  string `json:"key" validate:"required"`
-	Name string `json:"name"`
-}
-
-// NewFieldReference creates a new field reference with the given key and name
-func NewFieldReference(key string, name string) *FieldReference {
-	return &FieldReference{Key: key, Name: name}
-}
-
-// Type returns the name of the asset type
-func (r *FieldReference) Type() string {
-	return "field"
-}
-
-// Identity returns the unique identity of the asset
-func (r *FieldReference) Identity() string {
-	return string(r.Key)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *FieldReference) Variable() bool {
-	return false
-}
-
-func (r *FieldReference) String() string {
-	return fmt.Sprintf("%s[key=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ Reference = (*FieldReference)(nil)
-
-// FlowReference is used to reference a flow from another flow
-type FlowReference struct {
-	UUID FlowUUID `json:"uuid" validate:"required,uuid4"`
-	Name string   `json:"name"`
-}
-
-// NewFlowReference creates a new flow reference with the given UUID and name
-func NewFlowReference(uuid FlowUUID, name string) *FlowReference {
-	return &FlowReference{UUID: uuid, Name: name}
-}
-
-// Type returns the name of the asset type
-func (r *FlowReference) Type() string {
-	return "flow"
-}
-
-// GenericUUID returns the untyped UUID
-func (r *FlowReference) GenericUUID() uuids.UUID {
-	return uuids.UUID(r.UUID)
-}
-
-// Identity returns the unique identity of the asset
-func (r *FlowReference) Identity() string {
-	return string(r.UUID)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *FlowReference) Variable() bool {
-	return false
-}
-
-func (r *FlowReference) String() string {
-	return fmt.Sprintf("%s[uuid=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ UUIDReference = (*FlowReference)(nil)
-
-// GlobalReference is a reference to a global
-type GlobalReference struct {
-	Key  string `json:"key" validate:"required"`
-	Name string `json:"name"`
-}
-
-// NewGlobalReference creates a new global reference with the given key and name
-func NewGlobalReference(key string, name string) *GlobalReference {
-	return &GlobalReference{Key: key, Name: name}
-}
-
-// Type returns the name of the asset type
-func (r *GlobalReference) Type() string {
-	return "global"
-}
-
-// Identity returns the unique identity of the asset
-func (r *GlobalReference) Identity() string {
-	return string(r.Key)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *GlobalReference) Variable() bool {
-	return false
-}
-
-func (r *GlobalReference) String() string {
-	return fmt.Sprintf("%s[key=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ Reference = (*GlobalReference)(nil)
-
-// LabelReference is used to reference a label
-type LabelReference struct {
-	UUID      LabelUUID `json:"uuid,omitempty" validate:"omitempty,uuid4"`
-	Name      string    `json:"name,omitempty"`
-	NameMatch string    `json:"name_match,omitempty" engine:"evaluated"`
-}
-
-// NewLabelReference creates a new label reference with the given UUID and name
-func NewLabelReference(uuid LabelUUID, name string) *LabelReference {
-	return &LabelReference{UUID: uuid, Name: name}
-}
-
-// NewVariableLabelReference creates a new label reference from the given templatized name match
-func NewVariableLabelReference(nameMatch string) *LabelReference {
-	return &LabelReference{NameMatch: nameMatch}
-}
-
-// Type returns the name of the asset type
-func (r *LabelReference) Type() string {
-	return "label"
-}
-
-// GenericUUID returns the untyped UUID
-func (r *LabelReference) GenericUUID() uuids.UUID {
-	return uuids.UUID(r.UUID)
-}
-
-// Identity returns the unique identity of the asset
-func (r *LabelReference) Identity() string {
-	return string(r.UUID)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *LabelReference) Variable() bool {
-	return r.Identity() == ""
-}
-
-func (r *LabelReference) String() string {
-	return fmt.Sprintf("%s[uuid=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ UUIDReference = (*LabelReference)(nil)
-
-// TemplateReference is used to reference a Template
-type TemplateReference struct {
-	UUID TemplateUUID `json:"uuid" validate:"required,uuid"`
-	Name string       `json:"name"`
-}
-
-// NewTemplateReference creates a new template reference with the given UUID and name
-func NewTemplateReference(uuid TemplateUUID, name string) *TemplateReference {
-	return &TemplateReference{UUID: uuid, Name: name}
-}
-
-// GenericUUID returns the untyped UUID
-func (r *TemplateReference) GenericUUID() uuids.UUID {
-	return uuids.UUID(r.UUID)
-}
-
-// Identity returns the unique identity of the asset
-func (r *TemplateReference) Identity() string {
-	return string(r.UUID)
-}
-
-// Type returns the name of the asset type
-func (r *TemplateReference) Type() string {
-	return "template"
-}
-
-func (r *TemplateReference) String() string {
-	return fmt.Sprintf("%s[uuid=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *TemplateReference) Variable() bool {
-	return false
-}
-
-var _ UUIDReference = (*TemplateReference)(nil)
-
-// TicketerReference is used to reference a ticketer
-type TicketerReference struct {
-	UUID TicketerUUID `json:"uuid" validate:"required,uuid"`
-	Name string       `json:"name"`
-}
-
-// NewTicketerReference creates a new classifier reference with the given UUID and name
-func NewTicketerReference(uuid TicketerUUID, name string) *TicketerReference {
-	return &TicketerReference{UUID: uuid, Name: name}
-}
-
-// Type returns the name of the asset type
-func (r *TicketerReference) Type() string {
-	return "ticketer"
-}
-
-// GenericUUID returns the untyped UUID
-func (r *TicketerReference) GenericUUID() uuids.UUID {
-	return uuids.UUID(r.UUID)
-}
-
-// Identity returns the unique identity of the asset
-func (r *TicketerReference) Identity() string {
-	return string(r.UUID)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *TicketerReference) Variable() bool {
-	return false
-}
-
-func (r *TicketerReference) String() string {
-	return fmt.Sprintf("%s[uuid=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ UUIDReference = (*TicketerReference)(nil)
-
-// TopicReference is used to reference a topic
-type TopicReference struct {
-	UUID TopicUUID `json:"uuid" validate:"required,uuid"`
-	Name string    `json:"name"`
-}
-
-// NewTopicReference creates a new topic reference with the given UUID and name
-func NewTopicReference(uuid TopicUUID, name string) *TopicReference {
-	return &TopicReference{UUID: uuid, Name: name}
-}
-
-// Type returns the name of the asset type
-func (r *TopicReference) Type() string {
-	return "topic"
-}
-
-// GenericUUID returns the untyped UUID
-func (r *TopicReference) GenericUUID() uuids.UUID {
-	return uuids.UUID(r.UUID)
-}
-
-// Identity returns the unique identity of the asset
-func (r *TopicReference) Identity() string {
-	return string(r.UUID)
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *TopicReference) Variable() bool {
-	return false
-}
-
-func (r *TopicReference) String() string {
-	return fmt.Sprintf("%s[uuid=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-var _ UUIDReference = (*TopicReference)(nil)
-
-// UserReference is used to reference a user
-type UserReference struct {
-	Email string `json:"email" validate:"required"`
-	Name  string `json:"name"`
-}
-
-// NewUserReference creates a new user reference with the given key and name
-func NewUserReference(email, name string) *UserReference {
-	return &UserReference{Email: email, Name: name}
-}
-
-// Type returns the name of the asset type
-func (r *UserReference) Type() string {
-	return "user"
-}
-
-// Identity returns the unique identity of the asset
-func (r *UserReference) Identity() string {
-	return r.Email
-}
-
-// Variable returns whether this a variable (vs concrete) reference
-func (r *UserReference) Variable() bool {
-	return false
-}
-
-func (r *UserReference) String() string {
-	return fmt.Sprintf("%s[email=%s,name=%s]", r.Type(), r.Identity(), r.Name)
-}
-
-// UmarshalJSON unmarshals this object from JSON
-func (r *UserReference) UnmarshalJSON(data []byte) error {
-	// can be read from email string
-	if data[0] == '"' {
-		var email string
-		if err := jsonx.Unmarshal(data, &email); err != nil {
-			return err
-		}
-		r.Email = email
-		r.Name = ""
-		return nil
-	}
-
-	// or a JSON object with email/name properties
-	var raw map[string]string
-	if err := jsonx.Unmarshal(data, &raw); err != nil {
-		return err
-	}
-
-	r.Email = raw["email"]
-	r.Name = raw["name"]
-	return nil
-}
-
-var _ Reference = (*UserReference)(nil)
-
 //------------------------------------------------------------------------------------------
 // Callbacks for missing assets
 //------------------------------------------------------------------------------------------
@@ -470,30 +34,6 @@ var PanicOnMissing MissingCallback = func(a Reference, err error) { panic(fmt.Sp
 
 // IgnoreMissing does nothing if an asset is reported missing
 var IgnoreMissing MissingCallback = func(Reference, error) {}
-
-//------------------------------------------------------------------------------------------
-// Validation
-//------------------------------------------------------------------------------------------
-
-// GroupReferenceValidation validates that the given group reference is either a concrete
-// reference or a name matcher
-func GroupReferenceValidation(sl validator.StructLevel) {
-	ref := sl.Current().Interface().(GroupReference)
-	if neitherOrBoth(string(ref.UUID), ref.NameMatch) {
-		sl.ReportError(ref.UUID, "uuid", "UUID", "mutually_exclusive", "name_match")
-		sl.ReportError(ref.NameMatch, "name_match", "NameMatch", "mutually_exclusive", "uuid")
-	}
-}
-
-// LabelReferenceValidation validates that the given label reference is either a concrete
-// reference or a name matcher
-func LabelReferenceValidation(sl validator.StructLevel) {
-	ref := sl.Current().Interface().(LabelReference)
-	if neitherOrBoth(string(ref.UUID), ref.NameMatch) {
-		sl.ReportError(ref.UUID, "uuid", "UUID", "mutually_exclusive", "name_match")
-		sl.ReportError(ref.NameMatch, "name_match", "NameMatch", "mutually_exclusive", "uuid")
-	}
-}
 
 // utility method which returns true if both string values or neither string values is defined
 func neitherOrBoth(s1 string, s2 string) bool {
